+++ src/main/java/me/efekos/simpler/items/custom/FieldType.java
@@ -22,14 +22,11 @@
 
 package me.efekos.simpler.items.custom;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface SaveField {
-    String value();
-    FieldType fieldType() default FieldType.STRING;
+public enum FieldType {
+    STRING,
+    INTEGER,
+    DOUBLE,
+    LONG,
+    OBJECT,
+    ARRAY
 }